+++ ChangeLog.md
@@ -4,6 +4,8 @@
 
 - Replace minView with maxView in 'model search' mode
   - Higher ranges are more likely to produce models
+- Add test suite
+  - Uses PropaFP and 'Place' examples
 
 ## [v0.0.1](https://github.com/rasheedja/LPPaver/tree/v0.0.1)